requirements.txt
flask==2.*
google-auth>=2.23.0
google-api-python-client>=2.100.0
requests>=2.31.0
gunicorn==20.*
